refactor: Change like and comment count types to long for consistency
- Updated ProfilePostResponse, Post, and PostRepositoryCustom to use long instead of int for likeCount and commentCount, ensuring consistency across the application.
- Refactored PostService and PostRepositoryCustomImpl to accommodate the new data types in the retrieval of member posts.
- Adjusted unit tests to reflect the changes in response types and validate the updated functionality.

+++ src/main/java/com/example/instagramclone/domain/post/api/ProfilePostResponse.java
@@ -14,7 +14,7 @@ public record ProfilePostResponse(
         Long postId,
         String thumbnailUrl,
         boolean multipleImages,
-        int likeCount,
-        int commentCount
+        long likeCount,
+        long commentCount
 ) {
 }
